requirements.txt
pyserial
pandas
joblib
influxdb-client
scikit-learn
numpy